Isolate linear_bf16_fp32 ROCm variant into amd/ folder
The compressor's parity-aligned bf16->fp32 linear had the last remaining inline
if-hip in an NV kernel file. Move the ROCm variant (plain fp32 matmul, since
hipblas lacks bf16-in/fp32-out gemm) into a hand-maintained mirror under
miles_plugins/amd/models/deepseek_v4/, dispatched by a platform import fork in
compressor.py -- same idiom as qat.py / hyper_connection.py. The NV kernel file
reverts byte-for-byte to upstream. The mirror differs from NV by a single forward
line (annotated); its backward is a verbatim copy and must be kept in sync.

With this, all 5 DSv4 AMD kernels use one uniform "amd/ folder + import fork"
pattern and NV code stays byte-for-byte. Update JOURNEY isolation table (5 rows)
and converge the two TODOs (NV-GPU precision parity = must-do; sub-package = low
priority).

+++ miles_plugins/amd/models/deepseek_v4/precision_aligned_ops.py
@@ -17,7 +17,9 @@ class _BFloat16LinearFP32Func(torch.autograd.Function):
         ctx.weight_dtype = weight.dtype
 
         x_2d = x_bf16.reshape(-1, x_bf16.shape[-1])
-        out = torch.mm(x_2d, weight_bf16.t(), out_dtype=torch.float32)
+        # ROCm-only diff vs NV: hipblas has no bf16-in/fp32-out gemm (torch.mm out_dtype),
+        # so use a plain fp32 matmul (numerically >= cublas bf16-in/fp32-accumulate).
+        out = torch.mm(x_2d.float(), weight_bf16.t().float())
         return out.view(*x.shape[:-1], weight_bf16.shape[0])
 
     @staticmethod
